Remove unnecessary 'use_credentials' config field
The presence of credentials are governed by existence of a username. The
password can be left blank in order to be asked for at every connection.
So as long as there is a username, the credentials are enabled. Otherwise
they're not. So creating a bool field for it is redundant.